requirements.txt
python-dotenv==1.2.*
requests==2.32.*
terminaltables==3.1.*
